core_admin/requirements.txt
APScheduler==3.9.1
astropy==5.1.1
asttokens==2.1.0
backcall==0.2.0
bcrypt==4.0.1
cached-property==1.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
dill==0.3.6
Django==2.2.13
django-auth-ldap==4.1.0
django-cors-headers==2.5.2
django-filter==2.1.0
django-settings-export==1.2.1
django-url-filter==0.3.12
djangorestframework==3.9.2
entrypoints==0.4
enum-compat==0.0.3
executing==1.2.0
fonttools==4.38.0
globus-sdk==3.14.0
humanize==4.4.0
idna==3.4
ipykernel==6.17.1
ipyparallel==8.4.1
ipython==8.6.0
ipython-genutils==0.2.0
jedi==0.18.1
jupyter_client==7.4.5
jupyter_core==5.0.0
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
paramiko==2.12.0
parsl==1.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
prompt-toolkit==3.0.32
psutil==5.9.4
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyAesCrypt==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyerfa==2.0.0.1
Pygments==2.13.0
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-ldap==3.4.3
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
scipy==1.9.3
six==1.16.0
spiceypy==5.1.2
SQLAlchemy==1.3.2
sqlparse==0.4.3
stack-data==0.6.0
tblib==1.7.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typeguard==2.13.3
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
wcwidth==0.2.5
